Represent route tag with long to support uint32 max value

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/ConnectedRoute.java
@@ -85,7 +85,7 @@ public final class ConnectedRoute extends AbstractRoute {
   }
 
   @Override
-  public int getTag() {
+  public long getTag() {
     return NO_TAG;
   }
